handler.py
import runpod
import torch
import base64
import os
from io import BytesIO
from PIL import Image


# Импорты для Wan2.2 зависят от конкретной реализации (diffusers или кастомный код)
# Предположим использование стандартного пайплайна

def decode_base64_to_image(img_str):
    if "base64," in img_str:
        img_str = img_str.split("base64,")[1]
    img_data = base64.b64decode(img_str)
    return Image.open(BytesIO(img_data))


def handler(job):
    job_input = job['input']

    # Получаем параметры из твое.го бота
    prompt = job_input.get("prompt")
    image_b64 = job_input.get("image")
    num_frames = job_input.get("num_frames", 49)

    try:
        # 1. Подготовка изображения
        input_image = decode_base64_to_image(image_b64)

        # 2. Здесь логика вызова модели Wan2.2
        # Пример:
        # video_path = pipeline.generate(prompt, image=input_image, frames=num_frames)

        # 3. Загрузка результата (например, в S3 или возврат временной ссылки)
        # Для простоты вернем заглушку, которую твой бот ожидает:
        return {"video_url": "https://your-storage.com/output.mp4"}

    except Exception as e:
        return {"error": str(e)}


runpod.serverless.start({"handler": handler})