requirements.txt
runpod
torch
torchvision
diffusers
transformers
accelerate
pillow
sentencepiece